feat: add --default-branch-only cmdline option
Add a cmdline option to only mirror the default branch of a given action
between github.com and the target GitHub Enterprise instance, though
still mirroring all tags. This is particularly useful when syncing
verified publisher actions that may contain a large number of stale
and/or ongoing development branches and we only ever really want to sync
the tagged releases from the main branch for consumption. This
additionally avoids the "triggered actions storm" that can happen when
syncing many branches.

Contributes-to: actions/actions-sync#12

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>